API

API

Parameters.reconstruct — Method.

Make a new instance of a type with the same values as the input type except for the fields given in the associative second argument or as keywords.

type A; a; b end
a = A(3,4)
b = reconstruct(a, [(:b, 99)]) # ==A(3,99)

source

Parameters.type2dict — Method.

Transforms a type-instance into a dictionary.

julia> type T
           a
           b
       end

julia> type2dict(T(4,5))
Dict{Symbol,Any} with 2 entries:
  :a => 4
  :b => 5

source

Parameters.with_kw — Function.

This function is called by the @with_kw macro and does the syntax transformation from:

@with_kw immutable MM{R}
    r::R = 1000.
    a::R
end

into

immutable MM{R}
    r::R
    a::R
    MM(r,a) = new(r,a)
    MM(;r=1000., a=error("no default for a")) = MM{R}(r,a) # inner kw, type-paras are required when calling
end
MM{R}(r::R,a::R) = MM{R}(r,a) # default outer positional constructor
MM(;r=1000,a=error("no default for a")) =  MM(r,a) # outer kw, so no type-paras are needed when calling
MM(m::MM; kws...) = reconstruct(mm,kws)
MM(m::MM, di::Union{Associative, Tuple{Symbol,Any}}) = reconstruct(mm, di)
macro unpack_MM(varname)
    esc(quote
    r = varname.r
    a = varname.a
    end)
end
macro pack_MM(varname)
    esc(quote
    varname = Parameters.reconstruct(varname,r=r,a=a)
    end)
end

source

Parameters.@pack — Macro.

Packs variables into a composite type or a Dict{Symbol}

@pack dict_or_typeinstance = a, b, c

Example with dict:

a = 5.0
c = "Hi!"
d = Dict{Symbol,Any}()
@pack d = a, c
d # Dict{Symbol,Any}(:a=>5.0,:c=>"Hi!")

Example with type:

a = 99
c = "HaHa"
type A; a; b; c; end
d = A(4,7.0,"Hi")
@pack d = a, c
d.a == 99 #true
d.c == "HaHa" #true

source

Parameters.@unpack — Macro.

Unpacks fields/keys from a composite type or a Dict{Symbol} into variables

@unpack a, b, c = dict_or_typeinstance

Example with dict:

d = Dict{Symbol,Any}(:a=>5.0,:b=>2,:c=>"Hi!")
@unpack a, c = d
a == 5.0 #true
c == "Hi!" #true

Example with type:

type A; a; b; c; end
d = A(4,7.0,"Hi")
@unpack a, c = d
a == 4 #true
c == "Hi!" #true

source

Parameters.@with_kw — Macro.

Macro which allows default values for field types and a few other features.

Basic usage:

@with_kw immutable MM{R}
    r::R = 1000.
    a::Int = 4
end

For more details see manual.

source

Parameters.@with_kw_noshow — Macro.

As @with_kw but does not define a show method to avoid annoying redefinition warnings.

@with_kw_noshow immutable MM{R}
    r::R = 1000.
    a::Int = 4
end

For more details see manual.

source

Parameters.pack! — Function.

This function is invoked to pack one entity into some DataType and has signature:

pack!(dt::Any, ::Val{field}, value) -> value

Note that this means the only symbols or immutable field-descriptors are allowed, as they are used as type parameter in Val.

Two definitions are included in the package to pack into a composite type or into a dictionary with Symbol or string keys:

@inline pack!{f}(x, ::Val{f}, val) = setfield!(x, f, val)
@inline pack!{k}(x::Associative{Symbol}, ::Val{k}, val) = x[k]=val
@inline pack!{S<:AbstractString,k}(x::Associative{S}, ::Val{k}, val) = x[string(k)]=val

More methods can be added to allow for specialized packing of other datatypes.

See also unpack.

source

Parameters.unpack — Function.

This function is invoked to unpack one field/entry of some DataType dt and has signature:

unpack(dt::Any, ::Val{field}) -> value of field

The field is the symbol of the assigned variable.

Three definitions are included in the package to unpack a composite type or a dictionary with Symbol or string keys:

@inline unpack{f}(x, ::Val{f}) = getfield(x, f)
@inline unpack{k}(x::Associative{Symbol}, ::Val{k}) = x[k]
@inline unpack{S<:AbstractString,k}(x::Associative{S}, ::Val{k}) = x[string(k)]

More methods can be added to allow for specialized unpacking of other datatypes.

See also pack!.

source